Bootstrap the application
This adds all of the boilerplate stuff that we need for an Origami
service, as well as lots of TODO comments where we need to set up
external things.

Next step is to add in the basic unit and integration tests, then work
with the rest of the team to set up the Heroku apps and monitoring etc.