Add integration tests for main routes
This is more to demonstrate how integration tests can be written for the
HTML pages in this application. I'm asserting basic things like "there's
a link to each component on the home page" and "the component page
displays the requested version number".

This relies on a light mock Repo Data service, so that it can run
offline and the tests don't rely on the live Repo Data service.