Add data for services and image sets (#22)
Services and image sets now display all the information we have on them.
I've split things into partials a little more to help organise the
views. There are a few nasty things in the views at the moment, like
inline styles, but I think this is alright as we should be replacing all
the markup as part of the front end build.