Add cache-control headers (#36)
I'm caching most things for just 5 minutes now, we can increase if we
need to later. This will balance keeping origin traffic low but the site
will still update frequently.

I've also used stale-while-revalidate and stale-on-error headers so that
a user never has to wait for origin while the cache is warmed up.