Remove a conflicting attribute 🤦‍♂️
In today's episode of Rowan is an Idiot, he named one of the attributes
used by the component filter "data-o-component". If that looks familiar,
it's because that's the name of the attribute we use to auto-bind
components to DOM elements.

So what was happening is: each table row on the component page was being
initialised as the component it's supposed to be linking to. So that
o-toggle error we've been seeing, it's because we were initialising a
toggle on a table row which _of course_ doesn't have a target.

So much facepalm.