Add demo filtering to the sidebar (#123)
* Add demo filtering to the sidebar

This removes the component filter from the left hand side of a component
page, and now lists the demos using the same filter as the CodeDocs
pages.

I've renamed and moved the sidebar navigation partials to better-reflect
their new broader purpose. I also switched the whitespace of some views
to tabs for consistency, so it may be best to review this with
whitespace changes ignored.

* Reduce coverage expectation

It's bad that I'm doing this, but I didn't do anything that should have
reduced coverage. I'll investigate separately as this work is holding up
QA from being promoted.

* Nav bug and filter label.

* Restore demo page support info.

* Remove unnecessary class