Update default JsDoc node to display before a selection is made.
We should show classes or constructor functions named after
the component without a leading `o`. E.g the JS class for
`o-forms` if `Forms`. This pattern is largerly followed but
sometimes the leading `o` is present. E.g. the class for `o-table`
is `OTable`.

I don't believe it is in the spec but perhaps should be.
The leading `o` in component class names like `OTable` is the
exception to an otherwise quite consistent pattern.