Make component support messages friendlier
This updates all of the component support messages to be a little
friendlier. I've also allowed the "not maintained by Origami" and the
support level messages to appear alongside eachother. My logic for
this change is that if we want more non-Origami-maintained components
in the registry then users of those components still deserve to see
information about the support level.