Treat an origami type `module` the same as `component`.
Origami spec v2 drops module in favour of component.

I've added a note to the codebase "component.isComponent is confusing, "component"
could be any Origami project type, for example a library "lib". The `component`
variable should be refactored to `version` or `project` or similar. However
that impacts many views and related functions. I decided not to take that on,
since the notes add clarity and we may choose to replace repo data and thus
the registry logic in the short/mid term.

Repo data PR:
https://github.com/Financial-Times/origami-repo-data/pull/328

Conversation:
https://github.com/Financial-Times/origami-website/pull/273#discussion_r617396987