404 for any Origami component the Registry does not support explicitly.
We are very unlikely to present "O3" Origami components, components
with an `origamiVersion` set to `3.0` in the registry. Instead we
are using Storybook and intend to decommission the registry.

We also intend to decommission the Registry's source of data,
origami-repo-data, which will not ingest "O3" components.

In theory then, this change is not required. It's a little defensive.
However, if we change our minds on repo-data, or something goes wrong,
then the Registry should not behave in an unexpected way.

Note that this also fixes a bug related to `origamiVersion`. Repo
Data always returns a string but here we were assuming an integer.
We relied on mocks for our tests, which made the same incorrect
assumption.